Expanded Plant Iteration, Saved Temp Stats To Disk
Corrected row count in plant workbook iteration from 43 to 243.

Added try/catch block to the filter_temps_by_monitoring function to
account for the fact that not all sites have the same number of rows of
temperature data. The function now breaks when a datetime is found to
be None. A comment was added in the main function to explain this.

Temp stats are now recorded back into the output workbook before being
saved to disk.

Signed-off-by: Avery Uslaner <avery.uslaner@redbutte.utah.edu>